requirements.txt
Adafruit-Blinka==6.20.2
adafruit-circuitpython-busdevice==5.1.4
Adafruit-PlatformDetect==3.19.5
Adafruit-PureIO==1.1.9
board==1.0
circuitpython-nrf24l01==2.1.0
numpy==1.21.0
pkg_resources==0.0.0
pyftdi==0.53.3
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==4.3.2
RPi.GPIO==0.7.1
spidev==3.5
sysv-ipc==1.1.0
typing==3.7.4.3
